move per chain config
