+++ crates/proof/primitives/Cargo.toml
@@ -15,11 +15,15 @@ workspace = true
 async-trait.workspace = true
 thiserror.workspace = true
 alloy-primitives = { workspace = true }
+alloy-chains = { workspace = true }
+alloy-eips = { workspace = true }
 base-proof-preimage = { workspace = true }
+base-consensus-genesis = { workspace = true }
 serde = { workspace = true, optional = true, features = ["derive"] }
 jsonrpsee = { workspace = true, optional = true }
 
 [dev-dependencies]
+hex.workspace = true
 rstest.workspace = true
 serde_json = { workspace = true, features = ["std"] }
